Revert "Project management"
