+++ MiHRM/app/Helpers/Helpers.php
@@ -1,17 +1,13 @@
 <?php
 
 namespace App\Helpers;
-use Symfony\Component\HttpFoundation\Response;
 
 class Helpers
 {
     public static function result($message,$statusCode, $data=[]){
-
-        $statusText = Response::$statusTexts[$statusCode];
-        $statusCodeWithText ="{$statusCode}, {$statusText}";
         return response()->json([
             'message' => $message,
-            'status_code' => $statusCodeWithText,
+            'status_code' => $statusCode,
             'data' => $data
         ],$statusCode);
     }
