+++ MiHRM/app/Http/Controllers/AdminController.php
@@ -2,12 +2,12 @@
 
 namespace App\Http\Controllers;
 
-use App\Services\Admin\AdminService;
+use Illuminate\Http\Request;
+use App\Services\AdminService;
 use Illuminate\Http\JsonResponse;
-use App\Http\Requests\Admin\AssignProjectRequest;
-use App\Http\Requests\Admin\CreateProjectRequest;
-use App\Http\Requests\Admin\UpdateEmployeeRequest;
-use Illuminate\Http\Response;
+use App\Http\Requests\AssignProjectRequest;
+use App\Http\Requests\CreateProjectRequest;
+use App\Http\Requests\UpdateEmployeeRequest;
 
 class AdminController extends Controller
 {
@@ -18,14 +18,27 @@ class AdminController extends Controller
         $this->adminService = $adminService;
     }
 
-
+    /**
+     * Get employees by department ID
+     *
+     * @param int $department_id
+     * @return JsonResponse
+     */
     public function getEmployeesByDepartment($department_id): JsonResponse
     {
+        // Call the AdminService to fetch the employees and return response
         return $this->adminService->getEmployeesByDepartment($department_id);
     }
 
+    /**
+     * Delete a user and their related employee record
+     *
+     * @param int $user_id
+     * @return JsonResponse
+     */
     public function deleteUser($user_id): JsonResponse
     {
+        // Call the AdminService to delete the user and employee
         return $this->adminService->deleteUserAndEmployee($user_id);
     }
 
@@ -53,15 +66,13 @@ class AdminController extends Controller
 
     public function assignProject(AssignProjectRequest $request): JsonResponse
     {
+        // Pass the validated data to the service
         return $this->adminService->assignProject($request->validated());
     }
 
     public function getAllAssignedProjects(): JsonResponse
     {
+        // Delegate the task to AdminService
         return $this->adminService->getAllAssignedProjects();
-    }
-    public function showProjects()
-    {
-        return $this->adminService->getAllProjects();
     }
 }
