+++ MiHRM/app/Http/Controllers/EmployeeController.php
@@ -2,11 +2,9 @@
 
 namespace App\Http\Controllers;
 
-use App\Services\Employee\EmployeeService;
-use App\Http\Requests\Employee\LeaveRequest;
-use App\Http\Requests\UpdateProjectStatusRequest;
-use Illuminate\Http\JsonResponse;
-
+use App\Http\Requests\LeaveRequest;
+use App\Services\EmployeeService;
+use Illuminate\Http\Request;
 
 class EmployeeController extends Controller
 {
@@ -27,12 +25,8 @@ class EmployeeController extends Controller
 
     public function getAssignedProjects()
     {
+        // Delegate the task to EmployeeService
         return $this->employeeService->getAssignedProjects();
-    }
-    public function updateProjectStatus(UpdateProjectStatusRequest $request): JsonResponse
-    {
-       
-        return $this->employeeService->updateProjectStatus($request->validated());
     }
 
 }
