+++ MiHRM/app/Http/Middleware/LogRequest.php
@@ -2,8 +2,8 @@
 
 namespace App\Http\Middleware;
 
-use App\DTOs\LogsDTOs\RequestLogDTO;
 use App\Models\RequestLogs;
+use Illuminate\Support\Facades\Auth;
 use Closure;
 use Illuminate\Http\Request;
 
@@ -18,14 +18,19 @@ class LogRequest
      */
     public function handle(Request $request, Closure $next)
     {
-        $requestLogDTO = new RequestLogDTO($request);
-        $requestLog = RequestLogs::create($requestLogDTO->toArray());
+        $requestLog = RequestLogs::create([
+            'user_id' => Auth::check() ? Auth::id() : null,
+            'ip_address' => $request->ip(),
+            'user_agent' => $request->header('User-Agent'),
+            'request_method' => $request->method(),
+            'request_url' => $request->fullUrl(),
+            'request_body' => json_encode($request->all()),
+        ]);
         $request['request_log_id'] = $requestLog->id;
 
         // dd($requestLog);
        return $next($request);
     }
-
     public function terminate($request , $response){
         RequestLogs::find($request['request_log_id'])->update([
             'response_body' => json_encode($response->getContent()),
