+++ MiHRM/app/Http/Requests/BaseRequest.php
@@ -5,7 +5,6 @@ namespace App\Http\Requests;
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Contracts\Validation\Validator;
 use Illuminate\Http\Exceptions\HttpResponseException;
-use Illuminate\Http\Response;
 
 class BaseRequest extends FormRequest
 {
@@ -25,7 +24,7 @@ class BaseRequest extends FormRequest
         throw new HttpResponseException(response()->json([
             'message' => 'Validation failed',
             'errors' => $errors,
-        ], Response::HTTP_UNPROCESSABLE_ENTITY));
+        ], 422));
     }
 
     /**
