+++ MiHRM/app/Http/Requests/LeaveRequest.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace App\Http\Requests\Employee;
+namespace App\Http\Requests;
 
-use App\Http\Requests\BaseRequest;
+use Illuminate\Foundation\Http\FormRequest;
+
 class LeaveRequest extends BaseRequest
 {
     /**
