+++ MiHRM/app/Http/Requests/LoginRequest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Http\Requests\Auth;
+namespace App\Http\Requests;
 
-use App\Http\Requests\BaseRequest;
+use Illuminate\Foundation\Http\FormRequest;
 
 class LoginRequest extends BaseRequest
 {
