+++ MiHRM/app/Http/Requests/RegisterRequest.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace App\Http\Requests\Auth;
+namespace App\Http\Requests;
 
-use App\Http\Requests\BaseRequest;
+use Illuminate\Foundation\Http\FormRequest;
+
 class RegisterRequest extends BaseRequest
 {
     /**
