+++ MiHRM/app/Http/Requests/UpdateEmployeeRequest.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace App\Http\Requests\Admin;
+namespace App\Http\Requests;
 
-use App\Http\Requests\BaseRequest;
+use Illuminate\Foundation\Http\FormRequest;
+
 class UpdateEmployeeRequest extends BaseRequest
 {
     /**
@@ -12,7 +13,7 @@ class UpdateEmployeeRequest extends BaseRequest
      */
     public function authorize()
     {
-        return true;
+        return true; // Authorize the request for admin users
     }
 
     /**
@@ -23,8 +24,8 @@ class UpdateEmployeeRequest extends BaseRequest
     public function rules()
     {
         return [
-            'department_id' => 'required|exists:departments,id',
-            'position' => 'required|string|max:255',
+            'department_id' => 'required|exists:departments,id', // Check if department_id exists in the departments table
+            'position' => 'required|string|max:255', // Position is a required string field with max 255 characters
         ];
     }
