--- MiHRM/app/Http/Requests/UpdateProjectStatusRequest.php
@@ -1,31 +1,0 @@
-<?php
-
-namespace App\Http\Requests;
-
-use Illuminate\Foundation\Http\FormRequest;
-
-class UpdateProjectStatusRequest extends FormRequest
-{
-    /**
-     * Determine if the user is authorized to make this request.
-     *
-     * @return bool
-     */
-    public function authorize()
-    {
-        return true;
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array<string, mixed>
-     */
-    public function rules()
-    {
-        return [
-            'project_id' => 'required|integer|exists:projects,id',
-            'status'     => 'required|string|in:in_progress,pending,completed',
-        ];
-    }
-}
