--- MiHRM/app/Jobs/CronJobs/HandleLeave.php
@@ -1,58 +1,0 @@
-<?php
-
-namespace App\Jobs\CronJobs;
-
-use App\Models\Attendance;
-use App\Models\Employee;
-use App\Models\LeaveRequest;
-use Carbon\Carbon;
-use Illuminate\Bus\Queueable;
-use Illuminate\Contracts\Queue\ShouldBeUnique;
-use Illuminate\Contracts\Queue\ShouldQueue;
-use Illuminate\Foundation\Bus\Dispatchable;
-use Illuminate\Queue\InteractsWithQueue;
-use Illuminate\Queue\SerializesModels;
-
-class HandleLeave implements ShouldQueue
-{
-    use Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
-
-    /**
-     * Create a new job instance.
-     *
-     * @return void
-     */
-    public function __construct()
-    {
-        //
-    }
-
-    /**
-     * Execute the job.
-     *
-     * @return void
-     */
-    public function handle()
-{
-    $employees = Employee::all();
-    $today = Carbon::today()->toDateString();
-    // $today = '2024-10-11';
-    
-    foreach ($employees as $employee) {
-        $attendance = Attendance::firstOrCreate(
-            ['employee_id' => $employee->id, 'date' => $today],
-            ['status' => 'absent']
-        );
-
-        $leaveRequest = LeaveRequest::where('employee_id', $employee->id)
-            ->where('status', 'approved')
-            ->whereDate('start_date', '<=', $today)
-            ->whereDate('end_date', '>=', $today)
-            ->first();
-            
-        if ($leaveRequest) {
-            $attendance->update(['status' => 'onleave']);
-        }
-    }
-}
-}
