--- MiHRM/app/Models/ActivityLog.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-
-class ActivityLog extends Model
-{
-    use HasFactory;
-
-    protected $guarded = [];
-
-    public function user()
-    {
-        return $this->belongsTo(User::class);
-    }
-}
