+++ MiHRM/app/Models/Attendance.php
@@ -3,16 +3,13 @@
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Factories\HasFactory;
 
 class Attendance extends Model
 {
-    use HasFactory;
     protected $fillable = [
         'employee_id',
         'check_in_time',
-        'check_out_time', 
-        'working_hours', 
+        'check_out_time',  // Nullable field
         'date',
         'status',          // 'present', 'absent', 'onleave'
     ];
