+++ MiHRM/config/app.php
@@ -69,8 +69,7 @@ return [
     |
     */
 
-    'timezone' => 'Asia/Karachi',
-
+    'timezone' => 'UTC',
 
     /*
     |--------------------------------------------------------------------------
