--- MiHRM/database/seeders/AttendanceSeeder.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace Database\Seeders;
-
-use Illuminate\Database\Seeder;
-use App\Models\Attendance;
-
-class AttendanceSeeder extends Seeder
-{
-    /**
-     * Run the database seeds.
-     *
-     * @return void
-     */
-    public function run()
-    {
-        // Generate 50 attendance records
-        Attendance::factory()->count(50)->create(); // Now it will recognize factory method
-    }
-}
